refactor: fix canister name xrc
