+++ rs/src/canisters/mod.rs
@@ -2,4 +2,4 @@ pub mod cmc;
 pub mod governance;
 pub mod ledger;
 pub mod swap;
-pub mod xrd;
+pub mod xrc;
