+++ rs/src/canisters/xrc/constants.rs
@@ -1,3 +1,3 @@
-pub const XRD_CANISTER_ID: &str = "sgymv-uiaaa-aaaaa-aaaia-cai";
+pub const XRC_CANISTER_ID: &str = "sgymv-uiaaa-aaaaa-aaaia-cai";
 pub const QUOTE_USD: &str = "USD";
 pub const RATE_ICP: &str = "ICP";
