+++ rs/src/dashboard/types.rs
@@ -1,5 +1,5 @@
 pub mod state {
-    use crate::canisters::xrd::candid::ExchangeRate;
+    use crate::canisters::xrc::candid::ExchangeRate;
     use candid::CandidType;
     use serde::Deserialize;
     use std::collections::HashMap;
